Arreglos menores para que pase el porcentaje como float y no lo redondee

+++ src/Edificios/Reconstruible.java
@@ -4,7 +4,7 @@ public interface Reconstruible {
 
 		
 		public abstract void reconstruir(int puntosDeReconstruccion);
-		public abstract int porcentajeDeConstruccion();
+		public abstract float porcentajeDeConstruccion();
 		public abstract void actualizarPuntosDeConstruccion();
 	
 }
